requirements.txt
# Install latest versions
openai
python-dotenv
gradio
fastmcp
